+++ app/src/main/java/org/simple/clinic/instantsearch/InstantSearchEffect.kt
@@ -44,3 +44,7 @@ object ShowKeyboard : InstantSearchEffect()
 data class OpenShortCodeSearchScreen(val shortCode: String) : InstantSearchEffect()
 
 object OpenQrCodeScanner : InstantSearchEffect()
+
+data class CheckIfPatientAlreadyHasAnExistingNHID(val patientId: UUID) : InstantSearchEffect()
+
+object ShowNHIDErrorDialog : InstantSearchEffect()